fixbug/quran-reader
if audio is playing and we go to next page, then the audio src gets updated and selected aya is going to be played.
if audio is puased and we to go next page, only the src gets updated.

need to be implementd:
When audio src is getting of the server, and is not playing, we should show a loading icon for player.
when select a page of side menu, it doesn't work.